Make index.js an esm module and main.js the code.
